# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IpeSign API
  version: 0.2.0
  description: Single-use PDF signing and tamper-evident ledger API.
servers:
  - url: https://ipesign.onrender.com
  - url: http://localhost:8080
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ipesign_access
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Session:
      type: object
      required: [userId]
      properties:
        userId: {type: string}
        email: {type: string, format: email}
        displayName: {type: string}
        profileName: {type: string, description: Unverified presentation field.}
        ipeAddress: {type: string, description: Unverified presentation field.}
        role: {type: string}
    AuthInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 320}
        password: {type: string, format: password, maxLength: 1024}
    SignupInput:
      allOf:
        - $ref: "#/components/schemas/AuthInput"
        - type: object
          additionalProperties: false
          required: [name]
          properties:
            name: {type: string, minLength: 2, maxLength: 128}
            ipeAddress: {type: string, maxLength: 128}
    VerifyResult:
      type: object
      required: [valid, signatureValid, ledgerRecordValid, singleUseConfirmed]
      properties:
        valid: {type: boolean}
        signatureValid: {type: boolean}
        ledgerRecordValid: {type: boolean}
        singleUseConfirmed: {type: boolean}
        signerName: {type: string}
        signerEmail: {type: string}
        recordId: {type: string}
        certHash: {type: string}
paths:
  /v1/health:
    get:
      summary: Health and ledger summary
      responses:
        "200": {description: Healthy}
  /v1/ca:
    get:
      summary: Public certificate authority metadata
      responses:
        "200": {description: CA metadata}
  /v1/auth/config:
    get:
      summary: Return whether password authentication is enabled
      responses:
        "200": {description: Auth configuration}
  /v1/auth/signup:
    post:
      summary: Create a Supabase user through the backend
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SignupInput"}
      responses:
        "201":
          description: Signed in
          headers:
            Set-Cookie: {schema: {type: string}}
          content:
            application/json: {schema: {$ref: "#/components/schemas/Session"}}
        "202": {description: Email confirmation required}
        "400": {description: Invalid input, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /v1/auth/login:
    post:
      summary: Sign in through Supabase and set HttpOnly session cookies
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AuthInput"}
      responses:
        "200":
          description: Signed in
          headers:
            Set-Cookie: {schema: {type: string}}
          content:
            application/json: {schema: {$ref: "#/components/schemas/Session"}}
        "401": {description: Invalid credentials}
        "429": {description: Rate limited}
  /v1/auth/logout:
    post:
      summary: Revoke the current session and clear cookies
      responses:
        "204": {description: Signed out}
  /v1/auth/me:
    get:
      summary: Resolve or refresh the current session
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        "200":
          description: Session
          content:
            application/json: {schema: {$ref: "#/components/schemas/Session"}}
        "401": {description: Invalid or expired session}
  /v1/documents/sign:
    post:
      summary: Sign one PDF with a single-use certificate
      security: [{cookieAuth: []}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [pdf]
              properties:
                pdf: {type: string, format: binary, description: PDF up to 20 MiB.}
                policy_id: {type: string, maxLength: 128}
      responses:
        "200":
          description: Signed PDF containing the IpeSign envelope
          content:
            application/pdf: {schema: {type: string, format: binary}}
        "401": {description: Authentication required when Supabase is configured}
        "413": {description: Request body too large}
        "429": {description: Rate limited}
  /v1/documents/verify:
    post:
      summary: Verify a PDF containing an IpeSign envelope
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [pdf]
              properties:
                pdf: {type: string, format: binary, description: Signed PDF up to 20 MiB.}
      responses:
        "200":
          description: Verification result
          content:
            application/json: {schema: {$ref: "#/components/schemas/VerifyResult"}}
        "413": {description: Request body too large}
        "429": {description: Rate limited}
  /v1/records/{recordId}:
    get:
      summary: Look up and verify one ledger record
      parameters:
        - in: path
          name: recordId
          required: true
          schema: {type: string}
      responses:
        "200": {description: Verified record metadata}
        "404": {description: Record not found}
  /v1/chain/verify:
    get:
      summary: Verify the complete ledger hash chain and signatures
      responses:
        "200": {description: Ledger verification report}